Feat: implement modules functions
