+++ README.md
@@ -38,7 +38,7 @@ You should implement several functions in dedicated files
 
 You should refactor file (you can add additional imports if needed)
 
-- `cjsToEsm.cjs` - rewrite it to it's equivalent in ECMAScript notation (and rename it to `esm.mjs`)
+- `esm.mjs` - rewrite it to it's equivalent in ECMAScript notation (and rename it to `esm.mjs`)
 
 ### Hash (src/hash)
