+++ src/modules/files/c.mjs
@@ -1,0 +1,1 @@
+console.log('Hello from c.mjs!');
